+++ assets/Assets.php
@@ -19,13 +19,13 @@ class Assets extends AssetBundle
     public $sourcePath = '@calendar/resources';
 
     public $css = [
-        'js/fullcalendar/fullcalendar.min.css',
+        'js/fullcalendar/fullcalendar.css',
         'css/calendar.css',
         //'fullcalendar/fullcalendar.print.css', // print
     ];
     public $js = [
-        'js/fullcalendar/lib/moment.min.js',
-        'js/fullcalendar/fullcalendar.min.js',
+        'js/moment/moment.min.js',
+        'js/fullcalendar/fullcalendar.js',
         'js/fullcalendar/locale-all.js',
         'js/humhub.calendar.js'
     ];