+++ interfaces/CalendarItemsEvent.php
@@ -57,6 +57,13 @@ class CalendarItemsEvent extends CalendarEvent
         }
     }
 
+    public static function addItemTypes($itemTypesDef)
+    {
+        foreach ($itemTypesDef as $calendar) {
+            $event->addType($calendar->getItemTypeKey(), $calendar->getFullCalendarArray());
+        }
+    }
+
     /**
      * @inheritdoc
      */